refactor(callout): create new containers folder
