+++ markdown/src/translate/node.rs
@@ -1,9 +1,11 @@
 use std::fmt::Display;
 
-use pulldown_cmark::BlockQuoteKind;
 use serde::{Deserialize, Serialize};
 
-use super::element::{ElementTag, RenderElement};
+use super::{
+    container::callout::Callout,
+    element::{ElementTag, RenderElement},
+};
 
 // HTML
 
@@ -21,48 +23,6 @@ pub enum RenderIcon {
     Caution,
 }
 
-// CALLOUTS
-
-#[derive(Debug, Serialize, Deserialize)]
-pub enum CalloutKind {
-    Note,
-    Tip,
-    Important,
-    Warning,
-    Caution,
-}
-
-impl From<BlockQuoteKind> for CalloutKind {
-    fn from(value: BlockQuoteKind) -> Self {
-        match value {
-            BlockQuoteKind::Note => CalloutKind::Note,
-            BlockQuoteKind::Tip => CalloutKind::Tip,
-            BlockQuoteKind::Important => CalloutKind::Important,
-            BlockQuoteKind::Warning => CalloutKind::Warning,
-            BlockQuoteKind::Caution => CalloutKind::Caution,
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct RenderCallout {
-    pub kind: CalloutKind,
-    pub children: Vec<RenderNode>,
-}
-
-impl RenderCallout {
-    pub fn new(kind: CalloutKind) -> Self {
-        Self {
-            kind,
-            children: vec![],
-        }
-    }
-
-    pub fn add_child(&mut self, child: RenderNode) {
-        self.children.push(child)
-    }
-}
-
 // OVERARCHING NODE DATA TYPE
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -72,7 +32,7 @@ pub enum RenderNode {
     Icon(RenderIcon),
     Element(RenderElement),
     // Separate member because we want them to appear outside of <article>
-    Callout(RenderCallout),
+    Callout(Callout),
 }
 
 impl From<String> for RenderNode {
@@ -99,8 +59,8 @@ impl From<RenderElement> for RenderNode {
     }
 }
 
-impl From<RenderCallout> for RenderNode {
-    fn from(value: RenderCallout) -> Self {
+impl From<Callout> for RenderNode {
+    fn from(value: Callout) -> Self {
         RenderNode::Callout(value)
     }
 }
